Refactor : java convention
