+++ src/main/java/ladder/domain/Ladder.java
@@ -28,21 +28,22 @@ public class Ladder {
 		for (int i = 0; i < Integer.parseInt(ladderHegiht); i++) {
 			lines.add(Line.generateLine(countOfPerson));
 		}
+
 		return new Ladder(lines);
 	}
 
 	private static void validateCountOfPerson(int countOfPerson) {
-		if(countOfPerson < 2) {
+		if (countOfPerson < 2) {
 			throw new IllegalArgumentException(UserOutput.VIOLATE_NUMBER_OF_PLAYERS.getOutputMessage());
 		}
 	}
 
 	private static void validateLadderHeight(String ladderHeight) {
-		if(!ladderHeight.matches(UserOutput.REGEX_FOR_NUMBER.getOutputMessage())) {
+		if (!ladderHeight.matches(UserOutput.REGEX_FOR_NUMBER.getOutputMessage())) {
 			throw new IllegalArgumentException(UserOutput.VIOLATE_LADDER_HEIGHT.getOutputMessage());
 		}
 
-		if(Integer.parseInt(ladderHeight) <= 1) {
+		if (Integer.parseInt(ladderHeight) <= 1) {
 			throw new IllegalArgumentException(UserOutput.VIOLATE_LADDER_HEIGHT_LESS_THAN_TWO.getOutputMessage());
 		}
 	}
