+++ src/main/java/ladder/domain/LadderGameInformation.java
@@ -9,21 +9,21 @@ public class LadderGameInformation {
 	private final List<Reward> rewards;
 
 	public LadderGameInformation(List<Player> players, List<Reward> rewards) {
-		validateOverlapPlayers(players);
-		validatePlayersNumber(players);
+		validateOverlapPlayerNames(players);
+		validateCountOfPlayerNumber(players);
 		validateGameInformation(players, rewards);
 		this.players = players;
 		this.rewards = rewards;
 	}
 
-	private void validateOverlapPlayers(List<Player> players) {
+	private void validateOverlapPlayerNames(List<Player> players) {
 		if (new HashSet<>(players).size() != players.size()) {
 			throw new IllegalArgumentException(UserOutput.VIOLATE_PLAYER_OVERLAP_NAMES.getOutputMessage());
 		}
 	}
 
-	private void validatePlayersNumber(List<Player> players) {
-		if(players.size() < 2) {
+	private void validateCountOfPlayerNumber(List<Player> players) {
+		if (players.size() < 2) {
 			throw new IllegalArgumentException(UserOutput.VIOLATE_NUMBER_OF_PLAYERS.getOutputMessage());
 		}
 	}
