+++ src/main/java/ladder/domain/Reward.java
@@ -13,7 +13,7 @@ public class Reward {
 	}
 
 	private void validateRewardEmpty(String reward) {
-		if(StringUtils.isBlank(reward)) {
+		if (StringUtils.isBlank(reward)) {
 			throw new IllegalArgumentException(UserOutput.VIOLATE_GAME_REWARD.getOutputMessage());
 		}
 	}
