+++ src/test/java/ladder/domain/LadderTest.java
@@ -39,21 +39,21 @@ class LadderTest {
 
 	@Test
 	void 참가자의_수가_2미만인_경우_예외_반환() {
-		assertThrows(IllegalArgumentException.class, () ->  Ladder.generateLadder(1 ,"2"));
+		assertThrows(IllegalArgumentException.class, () -> Ladder.generateLadder(1, "2"));
 	}
 
 	@Test
 	void 참가자의_수가_2이상인_경우() {
-		assertDoesNotThrow(() ->  Ladder.generateLadder(2,"2"));
+		assertDoesNotThrow(() -> Ladder.generateLadder(2, "2"));
 	}
 
 	@Test
 	void 사다리의_높이가_2미만인_경우_예외_반환() {
-		assertThrows(IllegalArgumentException.class, () ->  Ladder.generateLadder(2,"1"));
+		assertThrows(IllegalArgumentException.class, () -> Ladder.generateLadder(2, "1"));
 	}
 
 	@Test
 	void 사다리의_높이가_2이상인_경우() {
-		assertDoesNotThrow(() ->  Ladder.generateLadder(2,"2"));
+		assertDoesNotThrow(() -> Ladder.generateLadder(2, "2"));
 	}
 }
