--- src/test/java/ladder/domain/LineTest.java
@@ -1,5 +1,0 @@
-package ladder.domain;
-
-public class LineTest {
-}
-
